pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-rag"
version = "0.2.0"
description = "Multi-agent Agentic RAG with Text-to-SQL and Vertex AI RAG Engine on ADK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "google-adk==1.5.0",
  "google-genai>=1.23.0",
  "google-cloud-aiplatform>=1.74.0",
  "google-cloud-logging>=3.11.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
]
mssql = [
  "pymssql>=2.2.8",
]
secrets = [
  "google-cloud-secret-manager>=2.20.0",
]
tenants = [
  "google-cloud-firestore>=2.18.0",
  "google-cloud-secret-manager>=2.20.0",
]
firebase = [
  "firebase-admin>=6.5.0",
]
presidio = [
  "presidio-analyzer>=2.2.0",
  "presidio-anonymizer>=2.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
